Integrated Convex to frontend applications
