+++ pnpm-workspace.yaml
@@ -13,6 +13,7 @@ catalog:
   react-dom: ^19.1.0
   typescript: ^5.9.2
   vitest: ^4.0.18
+  convex: ^1.31.7
 
 ignoredBuiltDependencies:
   - esbuild
