+++ server/convex/tasks.ts
@@ -1,11 +1,11 @@
-import { query, mutation } from "./_generated/server";
+import { query, mutation } from "./_generated/server.js";
 import { v } from "convex/values";
 
 export const sayHello = query({
-  args: {
-    name: v.string(),
-  },
-  handler: async (ctx, { name }) => {
-    return `Hello, ${name}!`;
-  },
+	args: {
+		name: v.string(),
+	},
+	handler: async (ctx, { name }) => {
+		return `Hello, ${name}!`;
+	},
 });
